pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wrapper"
version = "1.0.0"
description = "LLM and OCR provider integrations"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = []
keywords = ["llm", "ocr", "ai", "providers", "gemini", "azure", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "azure-ai-documentintelligence>=1.0.0",
    "aiohttp>=3.9.0",
    "filetype>=1.2.0",
    "Pillow>=10.0.0",
    "overrides>=7.7.0",
    "google-api-core>=2.15.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
llm = [
    "google-genai>=1.0.0",
    "openai>=1.0.0",
]
ocr = [
    "azure-ai-documentintelligence>=1.0.0",
]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "black>=24.1.1",
    "isort>=5.13.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
